Added a CLI for create_project.py and adjusted it

+++ annotator/create_project.py
@@ -10,30 +10,34 @@ def createProject(datasetID, dtd_file, xml_file, raw_data_folder, outdir='defaul
     
     project_dir = os.path.join(outdir, datasetID)
     
+    if not os.path.isfile(dtd_file):
+        raise FileNotFoundError(f"DTD file not found: {dtd_file}")
+    if not os.path.isfile(xml_file):
+        raise FileNotFoundError(f"XML file not found: {xml_file}")
+    if not os.path.isdir(raw_data_folder):
+        raise FileNotFoundError(f"Raw data folder not found: {raw_data_folder}")
+    
     if os.path.exists(project_dir):
-        
         if overwrite:
             shutil.rmtree(project_dir)
         else:
             msg = (f"A project with datasetID '{datasetID}' already exists at {project_dir}." +
-                   " Set createProject(overwrite=True) to overwrite the existing project.")
+                   " Set flag --overwrite to overwrite the existing project.")
             raise FileExistsError(msg)
     
     os.makedirs(project_dir, exist_ok=False)
     
     shutil.copy(xml_file, os.path.join(project_dir, f"{datasetID}.xml"))
     shutil.copy(dtd_file, os.path.join(project_dir, f"{datasetID}.dtd"))
-    
-    raw_data_dir = os.path.join(project_dir, f"raw_{datasetID}")
-    shutil.copytree(raw_data_folder, raw_data_dir)
+    shutil.copytree(raw_data_folder, os.path.join(project_dir, f"raw_{datasetID}"))
 
     print(f"Project <{datasetID}> created. Saved to:\n{project_dir}")
     
     return project_dir
     
-createProject("postcards",
-              r"C:\Users\tomas\Documents\Tese\AutoXML\data\postcards\postcards.dtd",
-              r"C:\Users\tomas\Documents\Tese\AutoXML\data\postcards\postcards.xml",
-              r"C:\Users\tomas\Documents\Tese\AutoXML\data\postcards\raw_postcards",
-              overwrite=True
-             )
+# createProject("postcards",
+              # r"C:\Users\tomas\Documents\Tese\AutoXML\data\postcards\postcards.dtd",
+              # r"C:\Users\tomas\Documents\Tese\AutoXML\data\postcards\postcards.xml",
+              # r"C:\Users\tomas\Documents\Tese\AutoXML\data\postcards\raw_postcards",
+              # overwrite=True
+             # )
